Move `Flume.slice_from_time()` to `.data._pathops` mod func
